+++ lib/glob/symbolize_keys.rb
@@ -6,7 +6,7 @@ module Glob
       case target
       when Hash
         target.each_with_object({}) do |(key, value), buffer|
-          buffer[key.to_sym] = SymbolizeKeys.call(value)
+          buffer[key.to_s.to_sym] = SymbolizeKeys.call(value)
         end
       when Array
         target.map {|item| SymbolizeKeys.call(item) }