+++ test/glob/symbolize_keys_test.rb
@@ -4,8 +4,8 @@ require "test_helper"
 
 class SymbolizeKeysTest < Minitest::Test
   test "symbolizes hash" do
-    input = {"a" => {"b" => 1}}
-    expected = {a: {b: 1}}
+    input = {"a" => {"b" => 1}, 42 => "the answer"}
+    expected = {a: {b: 1}, "42": "the answer"}
     actual = Glob::SymbolizeKeys.call(input)
 
     assert_equal expected, actual