+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Prefix your message with one of the following:
 - [Security] in case of vulnerabilities.
 -->
 
+## Unreleased
+
+- [Fixed] Properly handle numeric keys.
+
 ## v0.2.1 - 2022-01-13
 
 - [Changed] .gem package now include tests files.